feat(types): move user repository and document parser types to a different file
